requirements.txt
Flask
Gunicorn
flask-cors
python-dotenv
together
langchain
langchain-community
langchain-core
psycopg2-binary
pandas
openpyxl
torch
transformers
